Beatthemarket/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.1.5
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
channels==3.0.1
channels-redis==3.2.0
chardet==3.0.4
click==8.0.1
colorama==0.4.3
colorlover==0.3.0
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
daphne==3.0.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
django-bootstrap-static==4.2.1
django-cropperjs==0.0.5
django-letsencrypt==4.0.0
django-plotly-dash==1.6.4
django-tables2==2.4.0
django-timescaledb==0.2.8
dpd-components==0.1.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
feedparser==6.0.2
filelock==3.0.12
Flask==2.0.1
Flask-Compress==1.9.0
fpdf==1.7.2
futu-api==5.4.1608
future==0.18.2
gunicorn==20.1.0
hiredis==1.1.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
img2pdf==0.4.0
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.2
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mplfinance==0.12.7a7
msgpack==1.0.0
multitasking==0.0.9
numpy==1.19.1
# opencv-python==4.4.0.40
openpyxl==3.0.7
optional-django==0.1.0
packaging==20.4
pandas==1.1.4
pandas-datareader==0.9.0
parse==1.19.0
parso==0.7.1
pickleshare==0.7.5
pikepdf==2.8.0
Pillow==7.2.0
pipenv==2021.5.29
plotly==4.14.3
prompt-toolkit==3.0.8
protobuf==3.5.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyee==8.1.0
Pygments==2.7.2
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==20.0.0
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
retrying==1.3.3
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
toml==0.10.1
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
virtualenv==20.2.2
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
yahoo-fin==0.8.7
zipp==3.1.0
zope.interface==5.1.0
python-decouple==3.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0